# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Grocery Suggester API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface InventoryItem {
  id: number;
  name: string;
  category: string;
  quantity: number;
  unit: string;
  minThreshold: number;
  /** @nullable */
  expiresAt: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface InventoryItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  quantity: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  minThreshold: number;
  expiresAt?: string;
  notes?: string;
}

export interface InventoryItemUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  /** @minimum 0 */
  quantity?: number;
  unit?: string;
  /** @minimum 0 */
  minThreshold?: number;
  expiresAt?: string;
  notes?: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface InventoryStats {
  totalItems: number;
  lowStockCount: number;
  outOfStockCount: number;
  expiringCount: number;
  categoryBreakdown: CategoryCount[];
}

export type GrocerySuggestionUrgency = typeof GrocerySuggestionUrgency[keyof typeof GrocerySuggestionUrgency];


export const GrocerySuggestionUrgency = {
  critical: 'critical',
  low: 'low',
  expiring_soon: 'expiring_soon',
} as const;

export interface GrocerySuggestion {
  /** @nullable */
  inventoryItemId: number | null;
  name: string;
  category: string;
  currentQuantity: number;
  minThreshold: number;
  unit: string;
  reason: string;
  urgency: GrocerySuggestionUrgency;
}

export interface ShoppingList {
  id: number;
  name: string;
  itemCount: number;
  purchasedCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ShoppingListInput {
  /** @minLength 1 */
  name: string;
}

export interface ShoppingListItem {
  id: number;
  shoppingListId: number;
  /** @nullable */
  inventoryItemId: number | null;
  name: string;
  category: string;
  quantity: number;
  unit: string;
  purchased: boolean;
  /** @nullable */
  notes: string | null;
}

export interface ShoppingListWithItems {
  id: number;
  name: string;
  items: ShoppingListItem[];
  createdAt: string;
  updatedAt: string;
}

export interface ShoppingListItemInput {
  inventoryItemId?: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  quantity: number;
  /** @minLength 1 */
  unit: string;
  notes?: string;
}

export interface ShoppingListItemUpdate {
  purchased?: boolean;
  /** @minimum 0 */
  quantity?: number;
  notes?: string;
}

export interface PurchaseResult {
  purchasedCount: number;
  updatedInventoryItems: number;
}

export type ListInventoryItemsParams = {
category?: string;
lowStock?: boolean;
};