# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Grocery Suggester API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List all inventory items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: lowStock
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add a new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a single inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    InventoryItem:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        category:
          type: string
        quantity:
          type: number
        unit:
          type: string
        minThreshold:
          type: number
        expiresAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - quantity
        - unit
        - minThreshold
        - expiresAt
        - notes
        - createdAt
        - updatedAt
    InventoryItemInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        minThreshold:
          type: number
          minimum: 0
        expiresAt:
          type: string
        notes:
          type: string
      required:
        - name
        - category
        - quantity
        - unit
        - minThreshold
    InventoryItemUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
        minThreshold:
          type: number
          minimum: 0
        expiresAt:
          type: string
        notes:
          type: string
...