# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Grocery Suggester API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItem {
  id: number;
  name: string;
  category: string;
  quantity: number;
  unit: string;
  minThreshold: number;
  /** @nullable */
  expiresAt: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}