# lib/api-zod/src/generated/types/inventoryItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Grocery Suggester API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  quantity: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  minThreshold: number;
  expiresAt?: string;
  notes?: string;
}