# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Grocery Suggester API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  /** @minimum 0 */
  quantity?: number;
  unit?: string;
  /** @minimum 0 */
  minThreshold?: number;
  expiresAt?: string;
  notes?: string;
}